ssh接続及びログイン関係の処理をモジュール化
