+++ tex/list.tex
@@ -20,6 +20,10 @@
 \inputminted[linenos, breaklines]{rust}{src/cmdline_opt.rs}
 \clearpage
 
+\section{ssh2 ログイン処理モジュール ssh\_connect.rs}
+\inputminted[linenos, breaklines]{rust}{src/ssh_connect.rs}
+\clearpage
+
 \section{ファイルシステムモジュール ssh\_filesystem.rs}
 \inputminted[linenos, breaklines]{rust}{src/ssh_filesystem.rs}
